Revert "Merge branch 'sprint-1' into sign-up-page"
This reverts commit 875e99a2980ddcea7716f280a3854f74e59285ad, reversing
changes made to 87cc2b90978311e5a2c91d3435313d443e591a8b.
