+++ README.md
@@ -1,48 +1,1 @@
-# Band Geeks - News App
-
-## Team Members
-| Bonilla, Carlos D.                                                                                                                                             | Antonucci, Joseph                                                                                                                                                                                                                                                                                                                                                            | Pauliuchenka, Palina | Mukherjee, Samridhya | Ramirez, Sean A|
-|----------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------| ------------- | ------------- | ------------- |
-| <a href="https://github.com/cdb24" target="blank"><img src="https://img.shields.io/badge/GitHub-100000?style=for-the-badge&logo=github&logoColor=white" /></a> | <a href="https://www.linkedin.com/in/joseph-antonucci-23a823252/" target="blank"><img src="https://img.shields.io/badge/LinkedIn-0077B5?style=for-the-badge&logo=linkedin&logoColor=white" /></a><br/><a href="https://github.com/ChowChowSonic" target="blank"><img src="https://img.shields.io/badge/GitHub-100000?style=for-the-badge&logo=github&logoColor=white" /></a> | <a href="https://www.linkedin.com/in/ppauliuchenka02/" target="blank"><img src="https://img.shields.io/badge/LinkedIn-0077B5?style=for-the-badge&logo=linkedin&logoColor=white" /></a> <br/><a href="https://github.com/ppauliuchenka02" target="blank"><img src="https://img.shields.io/badge/GitHub-100000?style=for-the-badge&logo=github&logoColor=white" /></a> | <a href="https://github.com/SamridhyaM" target="blank"><img src="https://img.shields.io/badge/GitHub-100000?style=for-the-badge&logo=github&logoColor=white" /></a> | <a href="https://www.linkedin.com/in/seanaramirez/" target="blank"><img src="https://img.shields.io/badge/LinkedIn-0077B5?style=for-the-badge&logo=linkedin&logoColor=white" /></a><br/><a href="https://github.com/snaramirez872" target="blank"><img src="https://img.shields.io/badge/GitHub-100000?style=for-the-badge&logo=github&logoColor=white" /></a>
-
-<!-- | Software Engineer                                                                                                                                              | Software Engineer                                                                                                                                                                                                                                                                                                                                         | Software Engineer  | Software Engineer  | Software Engineer  | -->
-
-### Project Description
-The project is a personal news feed
-The news app will have the following capabilities
-
-- Obtain recent new headlines from https://newsapi.org
-- When no one is signed into the app, it will display articles from the “General” category
-- When a user is signed into the app, it will display articles from categories chosen in the user’s settings
-- There will be a page to sign up for the app
-- There will be a page to sign into the app
-- There will be a settings page for choosing new categories
-- There will be a search for users to search for specific topics
-
-### Tools used
-<div>
-    <p>Package & Module Managers:</p>
-    <img src="https://img.shields.io/badge/Webpack-8DD6F9?style=for-the-badge&logo=Webpack&logoColor=white" />
-    <img src="https://img.shields.io/badge/Yarn-2C8EBB?style=for-the-badge&logo=yarn&logoColor=white" />
-    <br />
-    <p>Frameworks, Queries & Libraries:</p>
-    <img src="https://img.shields.io/badge/Redwood-FDF8F6?style=for-the-badge&logo=redwoodjs&logoColor=BF4722" />
-    <img src="https://img.shields.io/badge/GraphQl-E10098?style=for-the-badge&logo=graphql&logoColor=white" />
-    <img src="https://img.shields.io/badge/Prisma-3982CE?style=for-the-badge&logo=Prisma&logoColor=white" />
-    <img src="https://img.shields.io/badge/React-20232A?style=for-the-badge&logo=react&logoColor=61DAFB" />
-    <br />
-    <img src="https://img.shields.io/badge/Tailwind_CSS-38B2AC?style=for-the-badge&logo=tailwind-css&logoColor=white" />
-    <br />
-    <p></p>API:</p>
-    <img height="30" src="https://camo.githubusercontent.com/52ea2d77e3c453c1cd1188cef5df627153bb420f8393b76c35559aaa3495ed5f/68747470733a2f2f692e696d6775722e636f6d2f68553367454e622e706e67" />
-    <img src="" />
-    <img src="" />
-</div>
-
-
-
-
-
-## Install & Run
-
-to be added later...
+# bandgeeks-news-app
