--- web/src/layouts/NewsLayout/NewsLayout.stories.js
@@ -1,10 +1,0 @@
-import NewsLayout from './NewsLayout'
-
-export const generated = (args) => {
-  return <NewsLayout {...args} />
-}
-
-export default {
-  title: 'Layouts/NewsLayout',
-  component: NewsLayout,
-}
