--- web/src/pages/AboutPage/AboutPage.js
@@ -1,13 +1,0 @@
-import { MetaTags } from '@redwoodjs/web'
-
-const AboutPage = () => {
-  return (
-    <>
-      <MetaTags title="About" description="About page" />
-
-      <main>About</main>
-    </>
-  )
-}
-
-export default AboutPage
