--- web/src/pages/AboutPage/AboutPage.stories.js
@@ -1,10 +1,0 @@
-import AboutPage from './AboutPage'
-
-export const generated = () => {
-  return <AboutPage />
-}
-
-export default {
-  title: 'Pages/AboutPage',
-  component: AboutPage,
-}
