--- web/src/pages/AboutPage/AboutPage.test.js
@@ -1,14 +1,0 @@
-import { render } from '@redwoodjs/testing/web'
-
-import AboutPage from './AboutPage'
-
-//   Improve this test with help from the Redwood Testing Doc:
-//   https://redwoodjs.com/docs/testing#testing-pages-layouts
-
-describe('AboutPage', () => {
-  it('renders successfully', () => {
-    expect(() => {
-      render(<AboutPage />)
-    }).not.toThrow()
-  })
-})
