--- web/src/pages/HomePage/HomePage.js
@@ -1,13 +1,0 @@
-import { MetaTags } from '@redwoodjs/web'
-
-const HomePage = () => {
-  return (
-    <>
-      <MetaTags title="Home" description="Home page" />
-
-      <main>Home</main>
-    </>
-  )
-}
-
-export default HomePage
