--- web/src/pages/HomePage/HomePage.stories.js
@@ -1,10 +1,0 @@
-import HomePage from './HomePage'
-
-export const generated = () => {
-  return <HomePage />
-}
-
-export default {
-  title: 'Pages/HomePage',
-  component: HomePage,
-}
