+++ web/src/pages/LandingPage/LandingPage.stories.js
@@ -1,0 +1,10 @@
+import LandingPage from './LandingPage'
+
+export const generated = () => {
+  return <LandingPage />
+}
+
+export default {
+  title: 'Pages/LandingPage',
+  component: LandingPage,
+}
