--- web/src/pages/MemePage/MemePage.js
@@ -1,16 +1,0 @@
-import { MetaTags } from '@redwoodjs/web'
-
-const MemePage = () => {
-  return (
-    <>
-      <MetaTags title="Meme" description="Meme page" />
-
-      <img
-        src="https://static1.srcdn.com/wordpress/wp-content/uploads/2020/08/Animal-Crossing-New-Horizons-Spongebob-Squarepants-Band-Geeks.jpg"
-        alt="band geeks"
-      />
-    </>
-  )
-}
-
-export default MemePage
