--- web/src/pages/MemePage/MemePage.stories.js
@@ -1,10 +1,0 @@
-import MemePage from './MemePage'
-
-export const generated = () => {
-  return <MemePage />
-}
-
-export default {
-  title: 'Pages/MemePage',
-  component: MemePage,
-}
